+++ lib/maps.h
@@ -6,6 +6,7 @@
 /* Defines for map properties */
 #define IMPASSABLE 0
 #define PASSABLE 1
+#define FRAME_SPEED 30
 
 /* Structure for map */
 struct worldmap {
@@ -17,6 +18,9 @@ struct worldmap {
 	unsigned short int **quantity;
 	unsigned short int **ground;
 	unsigned short int **roof;
+	char **frame;
+	unsigned short int **frame_count;
+	unsigned short int frame_speed;
 	SDL_bool **harvestable;
 };
 
@@ -27,6 +31,7 @@ extern short int	 get_sprite(int tile, short int biome);
 extern char		*get_color(int tile, short int biome);
 extern char		 is_passable(int tile, short int biome);
 extern char 		*get_tile_name(int tile, short int biome);
+extern void		 process_frames(struct worldmap *map, int row, int col);
 
 #endif